requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
colorama==0.4.6
conditional==1.4
coverage==7.2.1
crispy-bootstrap5==0.7
dal-admin-filters==1.1.0
dj-database-url==1.2.0
Django==4.1.7
django-admin-autocomplete-filter==0.7.1
django-appconf==1.0.5
-e git+https://github.com/yourlabs/django-autocomplete-light.git@62ecb5f2200d31c776ccac7bbd06c31c12be58f2#egg=django_autocomplete_light
django-bootstrap-modal-forms==2.2.1
django-clone==5.3.1
django-crispy-forms==2.0
django-glrm==1.1.3
django-picklefield==3.1
django-redis==5.2.0
django-select2==8.1.0
django-smart-selects==1.6.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djhacker==0.2.3
exceptiongroup==1.1.0
Faker==17.0.0
gunicorn==20.1.0
iniconfig==2.0.0
jsonlines==3.1.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pluggy==1.0.0
psycopg2==2.9.5
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
whitenoise==6.4.0
